Added changes from App-booking-engine-property-details, so it can be accepted by acceptance branch
