+++ app/Domits/src/features/bookingengine/simulateStripe.js
