+++ app/Domits/src/navigation/mainNavigationStack.js
@@ -7,14 +7,12 @@ import Scan from "../header/scan";
 import Pay from "../header/pay";
 import Bookings from "../header/bookings";
 import Pocket from "../header/pocket";
-import ListingDetailScreen from "../screens/listingDetailScreen/listingDetailScreen";
-import onBoarding1 from "../screens/bookingprocess/bookingProcess";
-import personalDetailsForm from "../screens/bookingprocess/personalDetailsForm";
-import finalBookingOverview from "../screens/bookingprocess/finalBookingOverview";
-import simulateStripe from "../screens/bookingprocess/simulateStripe";
-import paymentAccepted from "../screens/bookingprocess/paymentAccepted";
-import paymentDeclined from "../screens/bookingprocess/paymentDeclined";
-import bookedAccommodation from "../screens/bookingprocess/bookedAccommodation";
+import PropertyDetailsScreen from "../screens/propertyDetailsScreen/screens/propertyDetailsScreen";
+import BookingProcess from "../features/bookingengine/screens/BookingEngineScreen";
+import simulateStripe from "../features/bookingengine/simulateStripe";
+import paymentAccepted from "../features/bookingengine/screens/PaymentAcceptedScreen";
+import paymentDeclined from "../features/bookingengine/screens/PaymentDeclinedScreen";
+import guestBookings from "../screens/guestdashboard/GuestBookingsTab";
 import Landing from "../screens/Landing";
 import profile from "../screens/guestdashboard/GuestProfileTab";
 import paymentMethods from "../screens/guestdashboard/GuestPaymentsTab";
@@ -40,59 +38,48 @@ const Stack = createNativeStackNavigator();
 
 function MainNavigationStack() {
     return (
-        <Stack.Navigator
-            screenOptions={{
-                headerShown: false,
-            }}>
-            <Stack.Screen name="HomeScreen" component={HomeScreen} />
-            <Stack.Screen name="GuestDashboard" component={GuestDashboard} />
-            <Stack.Screen name="Scan" component={Scan} />
-            <Stack.Screen name="Pay" component={Pay} />
-            <Stack.Screen name="Bookings" component={Bookings} />
-            <Stack.Screen name="Pocket" component={Pocket} />
-            <Stack.Screen name="Detailpage" component={ListingDetailScreen} />
-            <Stack.Screen name="onBoarding1" component={onBoarding1} />
-            <Stack.Screen
-                name="personalDetailsForm"
-                component={personalDetailsForm}
-            />
-            <Stack.Screen
-                name="finalBookingOverview"
-                component={finalBookingOverview}
-            />
-            <Stack.Screen name="simulateStripe" component={simulateStripe} />
-            <Stack.Screen name="paymentAccepted" component={paymentAccepted} />
-            <Stack.Screen name="paymentDeclined" component={paymentDeclined} />
-            <Stack.Screen
-                name="bookedAccommodation"
-                component={bookedAccommodation}
-            />
-            <Stack.Screen name="Landing" component={Landing} />
-            <Stack.Screen name="Profile" component={profile} />
-            <Stack.Screen name="PaymentMethods" component={paymentMethods} />
-            <Stack.Screen name="Reviews" component={reviews} />
-            <Stack.Screen name="Settings" component={settings} />
-            <Stack.Screen name="HelpAndFeedback" component={helpAndFeedback} />
-            {/*<Stack.Screen name="HostHomepage" component={HostHomepage} />*/}
-            <Stack.Screen name="HostDashboard" component={HostDashboard} />
-            <Stack.Screen name="HostCalendar" component={HostCalendar} />
-            <Stack.Screen name="HostReviews" component={HostReviews} />
-            <Stack.Screen name="OnboardingHost" component={OnboardingHost} />
-            <Stack.Screen name="HostPayments" component={HostPayments} />
-            <Stack.Screen name="HostListings" component={HostListings} />
-            <Stack.Screen name="HostSettings" component={HostSettings} />
-            <Stack.Screen name="emailSettings" component={emailSettings} />
-            {/*<Stack.Screen name="HostRevenue" component={HostRevenue} />*/}
-            <Stack.Screen name="ListProperty" component={ListProperty} />
-            {/*<Stack.Screen name="LocationFillIn" component={LocationFillIn} />*/}
-            {/*<Stack.Screen name="selectAmenities" component={selectAmenities} />*/}
-            {/*<Stack.Screen name="PriceProperty" component={PriceProperty} />*/}
-            <Stack.Screen name="HostDetailPage" component={HostDetailPage} />
-            {/*<Stack.Screen*/}
-            {/*  name="ReviewAndSubmitScreen"*/}
-            {/*  component={ReviewAndSubmitScreen}*/}
-            {/*/>*/}
-        </Stack.Navigator>
+      <Stack.Navigator
+        screenOptions={{
+          headerShown: false,
+        }}>
+        <Stack.Screen name="HomeScreen" component={HomeScreen} />
+        <Stack.Screen name="GuestDashboard" component={GuestDashboard} />
+        <Stack.Screen name="Scan" component={Scan} />
+        <Stack.Screen name="Pay" component={Pay} />
+        <Stack.Screen name="Bookings" component={Bookings} />
+        <Stack.Screen name="Pocket" component={Pocket} />
+        <Stack.Screen name="Detailpage" component={PropertyDetailsScreen} />
+        <Stack.Screen name="BookingProcess" component={BookingProcess} />
+        <Stack.Screen name="simulateStripe" component={simulateStripe} />
+        <Stack.Screen name="paymentAccepted" component={paymentAccepted} />
+        <Stack.Screen name="paymentDeclined" component={paymentDeclined} />
+        <Stack.Screen name="guestBookings" component={guestBookings} />
+        <Stack.Screen name="Landing" component={Landing} />
+        <Stack.Screen name="Profile" component={profile} />
+        <Stack.Screen name="PaymentMethods" component={paymentMethods} />
+        <Stack.Screen name="Reviews" component={reviews} />
+        <Stack.Screen name="Settings" component={settings} />
+        <Stack.Screen name="HelpAndFeedback" component={helpAndFeedback} />
+        {/*<Stack.Screen name="HostHomepage" component={HostHomepage} />*/}
+        <Stack.Screen name="HostDashboard" component={HostDashboard} />
+        <Stack.Screen name="HostCalendar" component={HostCalendar} />
+        <Stack.Screen name="HostReviews" component={HostReviews} />
+        <Stack.Screen name="OnboardingHost" component={OnboardingHost} />
+        <Stack.Screen name="HostPayments" component={HostPayments} />
+        <Stack.Screen name="HostListings" component={HostListings} />
+        <Stack.Screen name="HostSettings" component={HostSettings} />
+        <Stack.Screen name="emailSettings" component={emailSettings} />
+        {/*<Stack.Screen name="HostRevenue" component={HostRevenue} />*/}
+        <Stack.Screen name="ListProperty" component={ListProperty} />
+        {/*<Stack.Screen name="LocationFillIn" component={LocationFillIn} />*/}
+        {/*<Stack.Screen name="selectAmenities" component={selectAmenities} />*/}
+        {/*<Stack.Screen name="PriceProperty" component={PriceProperty} />*/}
+        <Stack.Screen name="HostDetailPage" component={HostDetailPage} />
+        {/*<Stack.Screen*/}
+        {/*  name="ReviewAndSubmitScreen"*/}
+        {/*  component={ReviewAndSubmitScreen}*/}
+        {/*/>*/}
+      </Stack.Navigator>
     );
 }
